+++ CHANGELOG.md
@@ -8,7 +8,10 @@
   * fix: Added test cases for dimension key validation [#67](https://github.com/Floating-Dartists/matomo-tracker/pull/67)
   * refactor:
     * Removed `forcedId`
-    * Renamed `userId` to `uid`, renamed `TraceableClientMixin.traceName`, `TraceableWidget.traceName` and `trackScreenWithName.widgetName` to `actionName`
+    * Renamed `userId` to `uid`
+    * Renamed `TraceableClientMixin.traceName`, `TraceableWidget.traceName` and `trackScreenWithName.widgetName` to `actionName`
+    * Renamed `TraceableClientMixin.traceTitle` and `TraceableWidget.traceTitle` to `eventName`
+    * Renamed `TraceableClientMixin.widgetId` to `pvId`
 
 ## [3.1.0]